+++ Defs.hh
@@ -4,9 +4,9 @@
 #include <vector>
 
 
-// Declarations of multiple dimensions vectors of floats.
+// Declarations of multiple dimensions vectors of doubles.
 typedef std::vector<int> VI;
-typedef std::vector<float> VF;
+typedef std::vector<double> VF;
 typedef std::vector<VF> VVF;
 typedef std::vector<VVF> VVVF;
 typedef std::vector<VVVF> VVVVF;
